+++ api/routes/config.php
@@ -31,7 +31,6 @@ $app->patch('/config', function ($request, $response, $args) {
         $config = $ib->config->get();
         // Update the config values with the submitted data
         $ib->config->set($config, $data);
-        $ib->api->setAPIResponseMessage('Successfully updated configuration');
     }
 
 	$response->getBody()->write(jsonE($GLOBALS['api']));
@@ -42,13 +41,13 @@ $app->patch('/config', function ($request, $response, $args) {
 
 $app->get('/config/plugins', function ($request, $response, $args) {
 	$ib = ($request->getAttribute('ib')) ?? new ib();
-
-    $list = [];
-    foreach ($GLOBALS['plugins'] as $key => $value) {
-        $list[] = $value;
+    if ($ib->auth->checkAccess("ADMIN-CONFIG")) {
+        $list = [];
+        foreach ($GLOBALS['plugins'] as $key => $value) {
+            $list[] = $value;
+        }
+        $ib->api->setAPIResponseData($list);
     }
-	$ib->api->setAPIResponseData($list);
-
 	// Return the response
 	$response->getBody()->write(jsonE($GLOBALS['api']));
 	return $response
@@ -59,9 +58,9 @@ $app->get('/config/plugins', function ($request, $response, $args) {
 
 $app->get('/config/plugins/{plugin}', function ($request, $response, $args) {
 	$ib = ($request->getAttribute('ib')) ?? new ib();
-
-	$ib->api->setAPIResponseData($ib->config->get('Plugins',$args['plugin']));
-
+    if ($ib->auth->checkAccess("ADMIN-CONFIG")) {
+        $ib->api->setAPIResponseData($ib->config->get('Plugins',$args['plugin']));
+    }
 	// Return the response
 	$response->getBody()->write(jsonE($GLOBALS['api']));
 	return $response
@@ -76,7 +75,6 @@ $app->patch('/config/plugins/{plugin}', function ($request, $response, $args) {
         $config = $ib->config->get();
         // Update the config values with the submitted data
         $ib->config->setPlugin($config, $data, $args['plugin']);
-        $ib->api->setAPIResponseMessage('Successfully updated plugin configuration');
     }
 
 	$response->getBody()->write(jsonE($GLOBALS['api']));